[svn-r4155] Suck in all the PMI related functionality into the process manager component, instead of leaving it in the utilities. We are trying to make the process manager more general than just PMI.
Also, refactored the process manager code to dynamically pick what PMI
version we want to use, i.e., PMI-1 or PMI-2 (though the code for
PMI-2 is not in place yet). There are still some hacks because the PMI
client seems to be misbehaving with respect to the order in which it
sends init and initack; this needs to be investigated.