[svn-r4387] 1. Modified the proxy to return the exit status as is, instead of trying to convert it to the correct format (different UI implementations might want to do different things with the actual return code).
2. Modified the mpiexec UI to separate out return() or exit() kind of
return codes and signal-based terminations. In the case of
signal-based terminations, mpiexec just raises the same signal and
aborts. This allows the shell to find the correct string to print on
the screen. Plus, this allows debuggers to catch the signal and do a
trace dump as well.

This should fix ticket #518.