[svn-r5091] This fixes two parts in the process binding code:
1. When the system is hyper-threading enabled, PLPA does not give us
any direct indication of this. We estimate this based on the number of
processor IDs, number of socket IDs and number of core IDs. All
threads are assumed to be symmetric pair-wise.

2. We were not considering holes in the OS provided
processor/socket/core IDs. PLPA already provides support for this,
which we now take advantage of.

This should fix ticket #761.