[svn-r6122] Initial support for PMI tunneling.
The proxies trap PMI calls and multiplex them over a single connection
to the main PMI server. This reduces the number of fd resources used,
which is useful for large-scale systems. This can also potentially
allow the proxies to perform some intelligent tasks, such as caching,
in the future. This can possibly have a small performance penalty for
small scale systems, because of an additional forwarding, but should
be negligible.