[svn-r6281] Bug fix in the timeout or forced kill of a job.
Since we are now sharing the proxy control fd for PMI commands as well
as proxy control messages, it is possible that we end up in a cleanup
routine asynchronously while the main thread is working on a PMI
command. This results in data corruption in some cases for the data
being sent to the proxy.

This showed up as the KILL command not reaching the proxy correctly
causing the MPI jobs to never be terminated.