[svn-r6464] Bug fixes in the stdin code.
There were two problems:

1. We were never closing the stdin socket for the MPI process, so it
was not getting an EOF.

2. When multiple processes are forked, the socket in the proxy that
pointed to the stdin of the MPI process was being inherited by the
later forked processes as well. So, even if the proxy closes this
socket, since the other processes still had it open, the MPI process
was not getting an EOF.

Thanks to John Valdes for reporting this problem.