[svn-r6507] Bug fix in the PMI command reading code.
We were reading a large number of bytes and assuming that only one
command would be read at a time. This was OK for PMI-1, but with
multi-threading support in PMI-2, more than one command can come on
the same socket back-to-back, so the read buffer can contain multiple
PMI commands.