Cleanup multiple aspects of the is_local lookup.
1. Cleanup check for local IPs.  We check to see if the host IP
matches that of the list of local IP addresses.  If it does, it's a
local host.

2. Do not try to detect remotely accessible nodes.  Our logic for
identifying whether a node is remotely accessible is quite primitive
and hacky.  It doesn't serve any real purpose either.  This patch
deletes that code.

3. Cleanup the is_local lookup function.  If getifaddrs and inet_ntop
are available, try to detect local processes.  If they aren't there,
simply return non-local.  In this case, the launcher will fall back to
using a remote launch even for local processes.  Bad, but workable.

4. Make the hostname lookups more comprehensive, by matching the local
hostname and the IP address the local hostname resolves to as well.
This works well when the /etc/hosts file contains an IP address match
that is not a part of the network addresses.  In this case, the
hostname IP lookup will give that IP address, but it will not match
any of the local network IP addresses.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>