Check for number of arguments instead of a NULL argument.
Thanks to Masamichi Takagi @ NEC, Japan for reporting the error and
suggesting the patch.  The cached PUTs array was not reinitialized to
NULL values after the first time they were sent upstream and cleared.
So the next time we push them upstream, they contained garbage data
causing segfaults.  This patch looks for the number of arguments
rather than a NULL argument as a termination detector.

Signed-off-by: Masamichi Takagi <masamichi.takagi@gmail.com>