Create a dummy stdin socket instead of simply closing it.
For non rank-0 processes, store the stdin socket in a dummy variable
instead of passing NULL.  Passing NULL will cause the create_process
function to close the STDIN socket, allowing the process to reuse that
socket.  However, if an application reopens stdin, it causes an
incorrect socket (which is not STDIN) to be closed.  This is
technically a user application bug, but this is a safe-guard to
workaround that.  Fixes #1622.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>