Linker warning multiple common of 'HYD_pmcd_pmip'
The variable is declared as external in pmip.h, but defined in three
different files. The linker warns that it's actually merging them. If we
want to use an only variable declared as 'extern', it should be only
defined in one place. This commit removes two out of the three
declarations. See #1966.

Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>