Don't compress and order list of failed procs in PMI
Previously, PMI provided a list of failed processes as a sorted list via a
string. This meant the list could look something like this:

1,3-5,7,10

However, in the new fault tolerance specification, the function
MPI_COMM_FAILURE_ACK needs to be able to determine the local order of the
failures to more efficiently acknowledge them without creating a list per
communciator. This requires that PMI not sort or compress the failure
notification. So now, the previous string could look like this:

3,1,4,5,10,7

Obviously, this is less efficient if there are lots of failures. Hopefully,
this is something that can be fixed in future versions of PMI.

Signed-off-by: Junchao Zhang <jczhang@mcs.anl.gov>