allow timeout of zero in bscu_wait_for_completion
When hydra detects a timeout has occured, its method for killing
everything is to call HYDT_bsci_wait_for_completion with a timeout
of zero. However, the zero timeout value was ignored and processes
were not explicitly killed. Allowing for a value of zero makes timeout
more reliable and ensures mpiexec returns a non-zero status to the user.

Signed-off-by: Pavan Balaji <balaji@anl.gov>